modify android project
