+++ atip/android-test/test7.feature
@@ -1,13 +1,13 @@
 Feature: orientation & freeze & screenshot tests
     Scenario: functions
         When I launch "Settings" with "com.android.settings" and "Settings" on android
-        Then I press "home" key
+        Then I press "home" on android
         Then I take screenshot as "home.png"
-        And I sleep 3 seconds
+        Then I wait for 3 seconds
         Then I set orientation "n"
-        And I sleep 3 seconds
+        Then I wait for 3 seconds
         Then I open notification
-        And I sleep 3 seconds
+        Then I wait for 3 seconds
         Then I open quick settings
-        And I sleep 3 seconds
-        Then I press "home" key
+        Then I wait for 3 seconds
+        Then I press "home" on android
